Packets: Better packet registry dump and implement bundledump utility Packets: Additionally build all bundles as shared libraries

+++ senf/Packets/DefaultBundle/ICMPv6Packet.hh
@@ -47,7 +47,7 @@ namespace senf
     
     struct ICMPTypes {
         // ICMP type registry
-        typedef boost::uint16_t key_t;
+        typedef boost::uint8_t key_t;
     };
 
     /** \brief ICMP Packet for IPv6
